fix warnings

+++ src/tests/vpicio_mts.c
@@ -76,7 +76,7 @@ main(int argc, char **argv)
     uint64_t *offset_remote;
     uint64_t *mysize;
     double    t0, t1;
-    int       steps = 1, sleeptime = 0;
+    uint64_t  steps = 1, sleeptime = 0;
 
     pdcid_t transfer_request_x, transfer_request_y, transfer_request_z, transfer_request_px,
         transfer_request_py, transfer_request_pz, transfer_request_id1, transfer_request_id2;
@@ -199,7 +199,7 @@ main(int argc, char **argv)
         MPI_Barrier(MPI_COMM_WORLD);
         t0 = MPI_Wtime();
         if (rank == 0) {
-            printf("\n#Step  %llu\n", i);
+            printf("\n#Step  %lu\n", i);
         }
 #endif
         PDCprop_set_obj_time_step(obj_prop_xx, i);
